Fix compiler warnings (invert dependency between PublicClassXmlCommentFix and PublicClassXmlCommentAnalyzer)
